pyproject.toml
[tool.poetry]
name = "hotel_reservation_system"
version = "0.1.0"
description = ""
authors = ["Paweł Perenc <57562561+perqu@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
pre-commit = "^3.6.0"
click = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
